+++ packages/datastore/src/sync/index.ts
@@ -19,6 +19,7 @@ import {
 	SchemaModel,
 	SchemaNamespace,
 	TypeConstructorMap,
+	ModelPredicate,
 } from '../types';
 import { exhaustiveCheck, getNow, SYNC } from '../util';
 import DataStoreConnectivity from './datastoreConnectivity';
@@ -68,6 +69,7 @@ declare class ModelMetadata {
 	public readonly fullSyncInterval: number;
 	public readonly lastSync?: number;
 	public readonly lastFullSync?: number;
+	public readonly lastSyncPredicate?: null | string;
 }
 
 export enum ControlMessage {
@@ -102,7 +104,9 @@ export class SyncEngine {
 		private readonly maxRecordsToSync: number,
 		private readonly syncPageSize: number,
 		conflictHandler: ConflictHandler,
-		errorHandler: ErrorHandler
+		errorHandler: ErrorHandler,
+		private readonly syncPredicates: WeakMap<SchemaModel, ModelPredicate<any>>,
+		private readonly amplifyConfig: Record<string, any> = {}
 	) {
 		const MutationEvent = this.modelClasses[
 			'MutationEvent'
@@ -120,9 +124,14 @@ export class SyncEngine {
 		this.syncQueriesProcessor = new SyncProcessor(
 			this.schema,
 			this.maxRecordsToSync,
-			this.syncPageSize
+			this.syncPageSize,
+			this.syncPredicates
 		);
-		this.subscriptionsProcessor = new SubscriptionProcessor(this.schema);
+		this.subscriptionsProcessor = new SubscriptionProcessor(
+			this.schema,
+			this.syncPredicates,
+			this.amplifyConfig
+		);
 		this.mutationsProcessor = new MutationProcessor(
 			this.schema,
 			this.storage,
@@ -403,7 +412,14 @@ export class SyncEngine {
 	): Promise<Map<SchemaModel, [string, number]>> {
 		const modelLastSync: Map<SchemaModel, [string, number]> = new Map(
 			(await this.getModelsMetadata()).map(
-				({ namespace, model, lastSync, lastFullSync, fullSyncInterval }) => {
+				({
+					namespace,
+					model,
+					lastSync,
+					lastFullSync,
+					fullSyncInterval,
+					lastSyncPredicate,
+				}) => {
 					const nextFullSync = lastFullSync + fullSyncInterval;
 					const syncFrom =
 						!lastFullSync || nextFullSync < currentTimeStamp
@@ -687,38 +703,58 @@ export class SyncEngine {
 		const ModelMetadata = this.modelClasses
 			.ModelMetadata as PersistentModelConstructor<ModelMetadata>;
 
-		const models: [string, string][] = [];
+		const models: [string, SchemaModel][] = [];
+		let savedModel;
 
 		Object.values(this.schema.namespaces).forEach(namespace => {
 			Object.values(namespace.models)
 				.filter(({ syncable }) => syncable)
 				.forEach(model => {
-					models.push([namespace.name, model.name]);
+					models.push([namespace.name, model]);
 				});
 		});
 
 		const promises = models.map(async ([namespace, model]) => {
-			const modelMetadata = await this.getModelMetadata(namespace, model);
-			let savedModel: ModelMetadata;
+			const modelMetadata = await this.getModelMetadata(namespace, model.name);
+			const syncPredicate = ModelPredicateCreator.getPredicates(
+				this.syncPredicates.get(model),
+				false
+			);
+			const lastSyncPredicate = syncPredicate
+				? JSON.stringify(syncPredicate)
+				: null;
 
 			if (modelMetadata === undefined) {
 				[[savedModel]] = await this.storage.save(
 					this.modelInstanceCreator(ModelMetadata, {
-						model,
+						model: model.name,
 						namespace,
 						lastSync: null,
 						fullSyncInterval,
 						lastFullSync: null,
+						lastSyncPredicate,
 					}),
 					undefined,
 					ownSymbol
 				);
 			} else {
+				const prevSyncPredicate = modelMetadata.lastSyncPredicate
+					? modelMetadata.lastSyncPredicate
+					: null;
+				const syncPredicateUpdated = prevSyncPredicate !== lastSyncPredicate;
+
 				[[savedModel]] = await this.storage.save(
 					(this.modelClasses.ModelMetadata as PersistentModelConstructor<
 						any
 					>).copyOf(modelMetadata, draft => {
 						draft.fullSyncInterval = fullSyncInterval;
+						// perform a base sync if the syncPredicate changed in between calls to DataStore.start
+						// ensures that the local store contains all the data specified by the syncExpression
+						if (syncPredicateUpdated) {
+							draft.lastSync = null;
+							draft.lastFullSync = null;
+							draft.lastSyncPredicate = lastSyncPredicate;
+						}
 					})
 				);
 			}