+++ packages/predictions/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.13](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/predictions@3.2.12...@aws-amplify/predictions@3.2.13) (2020-11-23)
+
+**Note:** Version bump only for package @aws-amplify/predictions
+
+
+
+
+
+## [3.2.12](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/predictions@3.2.11...@aws-amplify/predictions@3.2.12) (2020-11-20)
+
+**Note:** Version bump only for package @aws-amplify/predictions
+
+
+
+
+
+## [3.2.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/predictions@3.2.10...@aws-amplify/predictions@3.2.11) (2020-11-13)
+
+**Note:** Version bump only for package @aws-amplify/predictions
+
+
+
+
+
+## [3.2.10](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/predictions@3.2.9...@aws-amplify/predictions@3.2.10) (2020-11-03)
+
+**Note:** Version bump only for package @aws-amplify/predictions
+
+
+
+
+
+## [3.2.9](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/predictions@3.2.8...@aws-amplify/predictions@3.2.9) (2020-10-31)
+
+**Note:** Version bump only for package @aws-amplify/predictions
+
+
+
+
+
+## [3.2.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/predictions@3.2.7...@aws-amplify/predictions@3.2.8) (2020-10-29)
+
+**Note:** Version bump only for package @aws-amplify/predictions
+
+
+
+
+
 ## [3.2.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/predictions@3.2.6...@aws-amplify/predictions@3.2.7) (2020-10-15)
 
 **Note:** Version bump only for package @aws-amplify/predictions